feat: added story; setup test and trying to get states working
