+++ starters/vue3-xstate-css/src/views/GreetView.vue
@@ -1,64 +1,7 @@
 <script setup lang="ts">
-import { createMachine, assign } from 'xstate';
 import { useMachine } from '@xstate/vue';
-import HeaderComponent from '../components/HeaderComponent.vue';
-
-const fetchGreeting = (message: string) => {
-  const encodedMessage = encodeURIComponent(message);
-  const endpoint = `https://api.starter.dev/.netlify/functions/server/hello?greeting=${encodedMessage}`;
-  return fetch(endpoint).then((result) => result.text());
-};
-
-const greetMachine = createMachine(
-  {
-    id: 'greet',
-    predictableActionArguments: true,
-    schema: {
-      context: {} as { query: string; message: string; error: string },
-      services: {} as {
-        callFetchMessage: {
-          data: { result: string };
-        };
-      },
-    },
-    initial: 'loading',
-    context: {
-      query: 'from This Dot Labs!',
-      message: '',
-      error: '',
-    },
-    states: {
-      loading: {
-        invoke: {
-          src: 'callFetchMessage',
-          onDone: {
-            target: 'complete',
-            actions: assign({
-              message: (_context, event) => event.data,
-            }),
-          },
-          onError: {
-            target: 'failure',
-            actions: assign({
-              error: (_context, event) => event.data.message,
-            }),
-          },
-        },
-      },
-      complete: {
-        type: 'final',
-      },
-      failure: {
-        type: 'final',
-      },
-    },
-  },
-  {
-    services: {
-      callFetchMessage: (context) => fetchGreeting(context.query),
-    },
-  }
-);
+import HeaderComponent from '@/components/HeaderComponent.vue';
+import { greetMachine } from '@/machines/greetMachine';
 
 const { state } = useMachine(greetMachine, { devTools: true });
 </script>
@@ -68,7 +11,7 @@ const { state } = useMachine(greetMachine, { devTools: true });
     <HeaderComponent> Fetch Data from API </HeaderComponent>
     <section class="fetch__section">
       <p class="fetch__section-title">Message:</p>
-      <div class="fetch__section-result">
+      <div class="fetch__section-result" data-cy="message-result">
         <div
           v-if="state.value === 'loading'"
           class="fetch__section-loader"
