+++ starters/vue3-xstate-css/src/views/__tests__/CounterView.cy.ts
@@ -1,4 +1,4 @@
-import CounterView from '../../views/CounterView.vue';
+import CounterView from '../CounterView.vue';
 
 describe('<CounterView />', () => {
   it('renders', () => {
